tests/requirements.txt
# Pytest and plugins
pytest>=7.0.0
pytest-timeout>=2.0.0
pytest-playwright>=0.4.0

# HTTP requests
requests>=2.28.0
urllib3>=1.26.0

# YAML parsing (for Helm chart validation)
PyYAML>=6.0

# Kubernetes client (optional, for advanced k8s operations)
kubernetes>=25.0.0

# AWS S3 client (for E2E cleanup)
boto3>=1.28.0

# NISE - OCP cost data generator (for E2E tests)
# Used by default for proper OCP data format
# Set E2E_USE_SIMPLE_DATA=true to skip NISE and use simplified format
koku-nise>=4.0.0